+++ vector/src/main/java/im/vector/app/features/crypto/verification/VerificationAction.kt
@@ -31,4 +31,5 @@ sealed class VerificationAction : VectorViewModelAction {
     object SkipVerification : VerificationAction()
     object VerifyFromPassphrase : VerificationAction()
     data class GotResultFromSsss(val cypherData: String, val alias: String) : VerificationAction()
+    object SecuredStorageHasBeenReset : VerificationAction()
 }