+++ CHANGES.md
@@ -8,6 +8,7 @@ Improvements 🙌:
  - Wording differentiation for direct rooms (#2176)
  - PIN code: request PIN code if phone has been locked
  - Small optimisation of scrolling experience in timeline (#2114)
+ - Allow user to reset cross signing if he has no way to recover (#2052)
 
 Bugfix 🐛:
  - Improve support for image/audio/video/file selection with intent changes (#1376)